Add files via upload
The register now page will be redirected as how other pages work when register now is clicked from the login page. The customization in the register page is that it pops a message whenever a field is left empty without the information it needs. Also if wrong characters are typed or if something else like @yahoo as such is missing, it indicates it too. Other than that, the login page is customized with blurred image at the background along with an EYE-CATCHING STYLE.